backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.30.0
redis==5.2.0
qdrant-client==1.12.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
openai==1.52.0
anthropic==0.39.0
edge-tts==6.1.12
faster-whisper==1.0.3
Pillow==11.0.0
jinja2==3.1.4
websockets==13.1
aiofiles==24.1.0
python-dotenv==1.0.1
